experiments/fcn_bnns/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
absl-py==2.0.0
    # via
    #   chex
    #   optax
    #   orbax-checkpoint
blackjax==1.0.0
    # via -r requirements.in
chex==0.1.85
    # via optax
etils[epath,epy]==1.5.2
    # via orbax-checkpoint
fastprogress==1.0.3
    # via blackjax
filelock==3.13.1
    # via
    #   torch
    #   triton
flax==0.7.5
    # via -r requirements.in
fsspec==2023.12.1
    # via
    #   etils
    #   torch
importlib-resources==6.1.1
    # via etils
jax==0.4.21
    # via
    #   -r requirements.in
    #   blackjax
    #   chex
    #   flax
    #   jaxopt
    #   numpyro
    #   optax
    #   orbax-checkpoint
jaxlib==0.4.21
    # via
    #   blackjax
    #   chex
    #   jaxopt
    #   numpyro
    #   optax
    #   orbax-checkpoint
jaxopt==0.8.2
    # via blackjax
jinja2==3.1.2
    # via torch
joblib==1.3.2
    # via -r requirements.in
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
ml-dtypes==0.3.1
    # via
    #   jax
    #   jaxlib
    #   tensorstore
mpmath==1.3.0
    # via sympy
msgpack==1.0.7
    # via
    #   flax
    #   orbax-checkpoint
multipledispatch==1.0.0
    # via numpyro
nest-asyncio==1.5.8
    # via orbax-checkpoint
networkx==3.2.1
    # via torch
numpy==1.26.2
    # via
    #   chex
    #   flax
    #   jax
    #   jaxlib
    #   jaxopt
    #   ml-dtypes
    #   numpyro
    #   opt-einsum
    #   optax
    #   orbax-checkpoint
    #   scipy
    #   tensorstore
numpyro==0.13.2
    # via -r requirements.in
nvidia-cublas-cu12==12.1.3.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.1.105
    # via torch
nvidia-cuda-nvrtc-cu12==12.1.105
    # via torch
nvidia-cuda-runtime-cu12==12.1.105
    # via torch
nvidia-cudnn-cu12==8.9.2.26
    # via torch
nvidia-cufft-cu12==11.0.2.54
    # via torch
nvidia-curand-cu12==10.3.2.106
    # via torch
nvidia-cusolver-cu12==11.4.5.107
    # via torch
nvidia-cusparse-cu12==12.1.0.106
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.19.3
    # via torch
nvidia-nvjitlink-cu12==12.3.101
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
nvidia-nvtx-cu12==12.1.105
    # via torch
opt-einsum==3.3.0
    # via jax
optax==0.1.7
    # via
    #   blackjax
    #   flax
orbax-checkpoint==0.4.5
    # via flax
protobuf==4.25.1
    # via orbax-checkpoint
pygments==2.17.2
    # via rich
pyyaml==6.0.1
    # via
    #   flax
    #   orbax-checkpoint
rich==13.7.0
    # via flax
scipy==1.11.4
    # via
    #   jax
    #   jaxlib
    #   jaxopt
sympy==1.12
    # via torch
tensorstore==0.1.51
    # via
    #   flax
    #   orbax-checkpoint
toolz==0.12.0
    # via chex
torch==2.2.0
    # via -r requirements.in
tqdm==4.66.1
    # via numpyro
triton==2.2.0
    # via torch
typing-extensions==4.8.0
    # via
    #   blackjax
    #   chex
    #   etils
    #   flax
    #   orbax-checkpoint
    #   torch
zipp==3.17.0
    # via etils
